# Improve compilation speed by taking advantage of memoization #77