--- a/include/boost/mp11/algorithm.hpp
+++ b/include/boost/mp11/algorithm.hpp
@@ -181,19 +181,19 @@ template<template<class...> class P, template<class...> class F> struct mp_trans
 
 #if BOOST_MP11_WORKAROUND( BOOST_MP11_MSVC, < 1920 )
 
-    template<class... U> struct _f_ { using type = mp_eval_if_q<mp_not<mp_invoke_q<Qp, U...>>, mp_first<mp_list<U...>>, Qf, U...>; };
-    template<class... U> using _f = typename _f_<U...>::type;
+    template<class... U> struct _f { using type = mp_eval_if_q<mp_not<mp_invoke_q<Qp, U...>>, mp_first<mp_list<U...>>, Qf, U...>; };
+    template<class... U> using fn = typename _f<U...>::type;
 
 #else
 
-    template<class... U> using _f = mp_eval_if_q<mp_not<mp_invoke_q<Qp, U...>>, mp_first<mp_list<U...>>, Qf, U...>;
+    template<class... U> using fn = mp_eval_if_q<mp_not<mp_invoke_q<Qp, U...>>, mp_first<mp_list<U...>>, Qf, U...>;
 
 #endif
 };
 
 template<template<class...> class P, template<class...> class F, class... L> struct mp_transform_if_impl
 {
-    using type = mp_transform<mp_transform_if_impl_f<P, F>::template _f, L...>;
+    using type = mp_transform_q<mp_transform_if_impl_f<P, F>, L...>;
 };
 
 } // namespace detail
@@ -209,12 +209,12 @@ template<template<class...> class P> struct mp_filter_impl_f
 {
     using Qp = mp_quote<P>;
 
-    template<class T1, class... T> using _f = mp_if< mp_invoke_q<Qp, T1, T...>, mp_list<T1>, mp_list<> >;
+    template<class T1, class... T> using fn = mp_if<mp_invoke_q<Qp, T1, T...>, mp_list<T1>, mp_list<> >;
 };
 
 template<template<class...> class P, class L1, class... L> struct mp_filter_impl
 {
-    using _t1 = mp_transform<mp_filter_impl_f<P>::template _f, L1, L...>;
+    using _t1 = mp_transform_q<mp_filter_impl_f<P>, L1, L...>;
     using _t2 = mp_apply<mp_append, _t1>;
 
     using type = mp_assign<L1, _t2>;
@@ -650,12 +650,12 @@ template<template<class...> class L, class T1, template<class...> class P> struc
 
 template<class T1, template<class...> class P> struct mp_sort_impl_f
 {
-    template<class U> using f = P<U, T1>;
+    template<class U> using fn = P<U, T1>;
 };
 
 template<template<class...> class L, class T1, class... T, template<class...> class P> struct mp_sort_impl<L<T1, T...>, P>
 {
-    using part = mp_partition<L<T...>, mp_sort_impl_f<T1, P>::template f>;
+    using part = mp_partition_q<L<T...>, mp_sort_impl_f<T1, P>>;
 
     using S1 = typename mp_sort_impl<mp_first<part>, P>::type;
     using S2 = typename mp_sort_impl<mp_second<part>, P>::type;
@@ -1046,12 +1046,12 @@ namespace detail
 
 template<class I> struct mp_replace_at_impl_p
 {
-    template<class T1, class T2> using _p = std::is_same<T2, I>;
+    template<class T1, class T2> using fn = std::is_same<T2, I>;
 };
 
 template<class W> struct mp_replace_at_impl_f
 {
-    template<class T1, class T2> using _f = W;
+    template<class T1, class T2> using fn = W;
 };
 
 #endif
@@ -1069,7 +1069,7 @@ template<class L, class I, class W> struct mp_replace_at_impl
 
 #else
 
-    using type = mp_transform_if<mp_replace_at_impl_p<mp_size_t<I::value>>::template _p, mp_replace_at_impl_f<W>::template _f, L, mp_iota<mp_size<L> > >;
+    using type = mp_transform_if_q<mp_replace_at_impl_p<mp_size_t<I::value>>, mp_replace_at_impl_f<W>, L, mp_iota<mp_size<L> > >;
 
 #endif
 };
@@ -1218,14 +1218,14 @@ template<template<class...> class L> struct mp_power_set_impl< L<> >
 
 template<class T1> struct mp_power_set_impl_f
 {
-    template<class L2> using _f = mp_push_front<L2, T1>;
+    template<class L2> using fn = mp_push_front<L2, T1>;
 };
 
 template<template<class...> class L, class T1, class... T> struct mp_power_set_impl< L<T1, T...> >
 {
     using S1 = mp_power_set< L<T...> >;
 
-    using S2 = mp_transform<mp_power_set_impl_f<T1>::template _f, S1>;
+    using S2 = mp_transform_q<mp_power_set_impl_f<T1>, S1>;
 
     using type = mp_append< S1, S2 >;
 };