--- a/include/boost/mp11/integer_sequence.hpp
+++ b/include/boost/mp11/integer_sequence.hpp
@@ -19,6 +19,8 @@
 #if defined(__has_builtin)
 # if __has_builtin(__make_integer_seq)
 #  define BOOST_MP11_HAS_MAKE_INTEGER_SEQ
+# elif __has_builtin(__integer_pack)
+#  define BOOST_MP11_HAS_INTEGER_PACK
 # endif
 #endif
 
@@ -35,6 +37,10 @@ template<class T, T... I> struct integer_sequence
 #if defined(BOOST_MP11_HAS_MAKE_INTEGER_SEQ)
 
 template<class T, T N> using make_integer_sequence = __make_integer_seq<integer_sequence, T, N>;
+
+#elif defined(BOOST_MP11_HAS_INTEGER_PACK)
+
+template<class T, T N> using make_integer_sequence = integer_sequence<T, __integer_pack(N)...>;
 
 #else