--- a/include/boost/mp11/map.hpp
+++ b/include/boost/mp11/map.hpp
@@ -38,11 +38,19 @@ template<template<class...> class M, class... U, class T> struct mp_map_replace_
 {
     using K = mp_first<T>;
 
-    // mp_replace_if is inlined here using a struct _f because of msvc-14.0
+#if BOOST_MP11_WORKAROUND( BOOST_MP11_MSVC, < 1920 )
 
     template<class V> struct _f { using type = mp_if< std::is_same<mp_first<V>, K>, T, V >; };
 
     using type = mp_if< mp_map_contains<M<U...>, K>, M<typename _f<U>::type...>, M<U..., T> >;
+
+#else
+
+    template<class V> using _f = mp_if< std::is_same<mp_first<V>, K>, T, V >;
+
+    using type = mp_if< mp_map_contains<M<U...>, K>, M<_f<U>...>, M<U..., T> >;
+
+#endif
 };
 
 } // namespace detail