@pdimov 2025-11-03
This looks good to me. I can only wonder why I failed to look at it two years ago when you submitted it originally; it probably came in an inopportune time when I was too busy.

I see a couple more things. It looks like you started to also optimize `mp_map_replace_impl` by moving the msvc workaround into an ifdef, but then didn't extract `_f` into a helper template.

Also, in the two cases where the `_impl` class is reduced to just having a single `using type = ...` (`mp_transform_if_impl` and `mp_map_erase_impl`), you can just remove the `_impl` class altogether and move the `...` part into the primary alias.

The 1.90 beta cutoff is the day after tomorrow, so I won't be able to merge this now, but to make sure I don't drop the ball again (sorry), please remind me after 1.90 is released.

And thank you for this work.