@jonathanpoelen 2025-11-08
> I see a couple more things. It looks like you started to also optimize `mp_map_replace_impl` by moving the msvc workaround into an ifdef, but then didn't extract `_f` into a helper template.

I went back to msvc because it wasn't working: boostorg/mp11@87719f69d2eee3ae1e4dd686c06507f77b14fead. However, after two years, the logs are no longer available.

> Also, in the two cases where the `_impl` class is reduced to just having a single `using type = ...` (`mp_transform_if_impl` and `mp_map_erase_impl`), you can just remove the `_impl` class altogether and move the `...` part into the primary alias.

Done.

> The 1.90 beta cutoff is the day after tomorrow, so I won't be able to merge this now, but to make sure I don't drop the ball again (sorry), please remind me after 1.90 is released.

I will try to remember :)