@jonathanpoelen 2025-11-09
I took the opportunity to go over what I had missed and use a gcc builtin for `make_integer_sequence`.

## `mp_nth_element`

compiler | gcc-15 | clang-20
-- | -- | --
before   | 0:00.66s - 206316K | 0:01.77s - 246832K
after    | 0:00.63s - 192944K | 0:01.76s - 245956K

```cpp
template<class I> using test
  = mp_nth_element<mp_iota_c<I::value+1>, I, mp_less>;

using r1 = mp_transform<test, mp_iota_c<30>>;
```

## `mp_map_find_impl` workaround (GCC)

compiler | gcc-15
-- | --
before   | 0:01.65s - 381364K
after    | 0:00.99s - 213064K

```cpp
template<class L, class M = mp_transform<mp_list, L>>
struct f { template<class I> using g = mp_map_find<M, I>; };

template<class I, class L = mp_iota<I>> using test
  = mp_transform<f<L>::template g, L>;

using r1 = mp_transform<test, mp_iota_c<50>>;
```

## `mp_transform` with 5 lists or more

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.21s - 80328K | 0:00.29s - 117756K
after    | 0:00.20s - 77408K | 0:00.29s - 117468K

```cpp
template<class I, class L = mp_iota<I>> using test
  = mp_transform<mp_list, L, L, L, L, L, L, L, L>;

using r1 = mp_transform<test, mp_iota_c<50>>;
```

## `mp_fill`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.15s - 60740K | 0:00.14s - 94144K
after    | 0:00.10s - 43504K | 0:00.14s - 94188K

```cpp
template<class I, class L = mp_iota<I>> using test
  = mp_fill<L, int>;

using r1 = mp_transform<test, mp_iota_c<100>>;
```

## `mp_replace_if`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.23s - 76088K | 0:00.37s - 116984K
after    | 0:00.22s - 75348K | 0:00.35s - 114952K

```cpp
template<class> using pred = mp_false;
template<class I, class L = mp_iota<I>> using test
  = mp_replace_if<L, pred, I>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `mp_replace`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.30s - 90020K | 0:00.48s - 129272K
after    | 0:00.29s - 89428K | 0:00.46s - 126580K

```cpp
template<class I, class L = mp_iota<I>> using test
  = mp_replace<L, I, I>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `mp_remove`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.32s - 91276K | 0:00.69s - 156512K
after    | 0:00.31s - 90684K | 0:00.67s - 157560K

```cpp
template<class I, class L = mp_iota<I>> using test
  = mp_remove<L, I>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `mp_sliding_fold`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.23s - 79084K | 0:00.48s - 135100K
after    | 0:00.22s - 77748K | 0:00.46s - 132436K

```cpp
template<class I, class L = mp_iota<I>> using test
  = mp_list<
      mp_sliding_fold<L, mp_int<2>, mp_plus>,
      mp_sliding_fold<L, mp_int<2>, mp_max>
    >;

using r1 = mp_transform<test, mp_iota_c<50>>;
```

## `mp_copy_if`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.21s - 65820K | 0:00.51s - 138048K
after    | 0:00.21s - 65428K | 0:00.49s - 138540K

```cpp
template<class> using pred = mp_false;
template<class I, class L = mp_iota<I>> using test
  = mp_copy_if<L, pred>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `mp_remove_if`

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.21s - 67308K | 0:00.51s - 137744K
after    | 0:00.21s - 66956K | 0:00.49s - 138492K

```cpp
template<class> using pred = mp_true;
template<class I, class L = mp_iota<I>> using test
  = mp_remove_if<L, pred>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `mp_map_replace`

compiler |      gcc-15        |     clang-20
-- | -- | --
before   | 0:00.39s - 114524K | 0:00.66s - 155888K
after    | 0:00.38s - 114140K | 0:00.64s - 155184K

```cpp
template<class I, class L = mp_transform<mp_list, mp_iota<I>>> using test
  = mp_map_replace<L, mp_list<I,I>>;

using r1 = mp_transform<test, mp_iota_c<120>>;
```

## `integer_sequence` with gcc

Use the `__integer_pack` builtin

compiler |      gcc-15
-- | --
before   | 0:00.14s - 47284K
after    | 0:00.07s - 37572K

```cpp
template<class I> using test
  = make_integer_sequence<int, I::value>;

using r1 = mp_transform<test, mp_iota_c<200>>;
```

## `integer_sequence` when no builtin is available

The probability that this code will be used is very low since msvc and clang-based compilers use `__make_integer_seq` and gcc uses `__integer_pack`. But it is shorter and eliminates the concatenation phase.

when `T` is other than `std::size_t` (here `int`):

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.14s - 47220K | 0:00.23s - 107632K
after    | 0:00.12s - 42344K | 0:00.23s - 107344K

when `T` is `std::size_t`:

compiler |      gcc-15       |     clang-20
-- | -- | --
before   | 0:00.13s - 46600K | 0:00.22s - 107228K
after    | 0:00.09s - 39808K | 0:00.15s - 94860K

```cpp
using T = int;

template<class I> using test
  = make_integer_sequence<T, I::value>;

using r1 = mp_transform<test, mp_iota_c<200>>;
```